Download an ICOS digital object

Download a digital object identified by a persistent identifier.

icos_download(dobj, path = tempdir(), internal = FALSE, verbose = TRUE)

Arguments

dobj
digital object identified by a persistent identifier (PID) URI

path
where to save the downloaded data (default = tempdir())

internal
read in data and return an internal data frame to the current workspace

verbose
verbose feeback on licensing, needs you to acknowledge the license agreement for a download - setting this to FALSE overrides this but by doing so you acknowledge the statement in bulk!

Value

a downloaded digital file on disk for further processing

Examples

if (FALSE) {
dobj <- "https://meta.icos-cp.eu/objects/lNJPHqvsMuTAh-3DOvJejgYc"
icos_download(dobj = dobj, verbose = FALSE)
}